Use simplified molecule/atom/bond objects to exchange with OpenBabel
TODO appropriately use scaling parameter (bond length), especially for Wikidata query

Background: The Windows version fails when trying to communicate with
Openbabel, since obabeliface is apparently not running in the same "environment"
(not sure what precisely), so that it does not get the static pointer to the
main QApplication. This then causes problems because the Atom object
tries to get some default fonts. Generally, these dependencies should be
disentangled a bit.