Clean up MolScene properly
This is probably not the final word on this: In Qt6, AtomPopupUnitTest.testCoordinateTransferChangeFromScene
crashed because the QGraphicsScene destructor would delete all items of the scene,
which led to the QGraphicsItem destructor of the atom being called,
which in turn appears to call some signal on the scene, which at this point
is no longer recognized as that object type by a cast (the stacktrace
was not complete and neither do I have all the Qt source code at hand).
So just to be on the safe side, I clear the scene in the MolScene destructor
(note that this needs to be done _before_ deleting the impl pointer
as the private data object is exchanged as part of clear()).