Adapt tests for Qt6
For some reason, the font metrics in Qt6 seem to be somewhat different
from Qt5. I yet have to understand the problem completely. But the tests
that check precise sizes of bounding boxes or SVG coordinates should
overall be re-vamped anyhow in order to not critically depend on a set
of font metrics.

Also: the performance for inserting a complex molecule has taken a significant hit.