Disable/adapt test for radical electrons
Currently, we just take the color of a radical electron from the parent.
In the GUI, there is no way of setting the color explicitly and in XML, there is currently no way to mark the color as invalid.
Consequently, all invalid (default-initialized) colors would be read back as plain black, which would be odd if the color of the atom is different.